yatube_api/api/permissions.py
from rest_framework import permissions


class IsAuthor(permissions.BasePermission):

    def has_object_permission(self, request, view, obj):
        if request.method in permissions.SAFE_METHODS:
            return super().has_permission(request, view)
        return obj.author == request.user
